pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
discord-py-interactions = "^5.13.2"
tortoise-orm = { extras = ["asyncpg"], version = "^0.23.0" }
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
